+++ benchmarks/benchmark_flash_attention.py
@@ -6,8 +6,8 @@ from einops import rearrange, repeat
 from torch import nn
 from torch.nn import functional as F
 
-from flash_attn.bert_padding import pad_input, unpad_input
-from flash_attn.flash_attn_interface import flash_attn_unpadded_qkvpacked_func
+from flash_attn.functional.bert_padding import pad_input, unpad_input
+from flash_attn.functional.flash_attn_interface import flash_attn_unpadded_qkvpacked_func
 from flash_attn.utils.benchmark import benchmark_all, benchmark_backward, benchmark_combined, benchmark_forward
