+++ flash_attn/functional/fused_dense.py
@@ -13,7 +13,7 @@ from torch.cuda.amp import custom_bwd, custom_fwd
 from torch.distributed import ProcessGroup
 from torch.nn import functional as F
 
-from flash_attn.ops.activations import gelu_bwd, relu_bwd, sqrelu_bwd, sqrelu_fwd
+from flash_attn.functional.activations import gelu_bwd, relu_bwd, sqrelu_bwd, sqrelu_fwd
 from flash_attn.utils.distributed import all_gather_raw, all_reduce, all_reduce_raw, reduce_scatter, reduce_scatter_raw
