+++ tests/functional/test_fused_dense.py
