+++ tests/functional/test_fused_dense_parallel.py
