+++ tests/modules/test_cross_entropy.py
