+++ tests/modules/test_cross_entropy_parallel.py
